jimt/dance-ai-modal-refactoring - moved ai-types back to types
