+++ apps/src/dance/ai/DanceAiScore.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import moduleStyles from './dance-ai-score.module.scss';
 import classNames from 'classnames';
-import {GeneratedEffectScores, MinMax} from './ai-types';
+import {GeneratedEffectScores, MinMax} from './types';
 
 export enum ScoreColors {
   // A grey bar with no symbol in the symbol container yet.
